+++ nodalia-cards.manifest.js
@@ -1,10 +1,10 @@
 export default {
-  "pkgVersion": "1.1.1-alpha.3",
-  "contentSha256_12": "6fcdc4a96d04",
+  "pkgVersion": "1.1.1-alpha.4",
+  "contentSha256_12": "4bbcdcf72f7e",
   "file": "nodalia-cards.bundle.js",
   "loaderFile": "nodalia-cards.js",
-  "hacsFile": "nodalia-cards-1.1.1-alpha.3.js"
+  "hacsFile": "nodalia-cards-1.1.1-alpha.4.js"
 };
-export const pkgVersion = "1.1.1-alpha.3";
-export const contentSha256_12 = "6fcdc4a96d04";
+export const pkgVersion = "1.1.1-alpha.4";
+export const contentSha256_12 = "4bbcdcf72f7e";
 export const file = "nodalia-cards.bundle.js";